feat(pishrink): restrict SD card expansion to cross-manufacturer support
If raspi-config is the mechanism to grow the image partition on
boot, modify the underlying configuration so that it won't fill
the full card. SD Cards can be different sizes and raw clones will
not be compatible if fully expanded file system. 92% is safe enough
to ensure compatibility across 32GB cards in 2022.